google-pubsub-perf/requirements.txt
docopt==0.6.2
google-cloud-pubsub==1.2.0
numpy==1.18.1
